changing PID params from int to float

+++ include/pid.h
@@ -1,3 +1,3 @@
-void setPIDconfs(int setpoint, int min, int max, int Kp, int Ki, int Kd);
+void setPIDconfs(int setpoint, int min, int max, float Kp, float Ki, float Kd);
 void updatePID(double val);
 double getPIDoutput();
